refactor(components/containers): update imports from index
