+++ client/src/components/Dashboard/UserLesson/UserLesson.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { IUserLessonProps } from '../../../interfaces/componentProps';
+import { IUserLessonProps } from '../../../interfaces';
 
 export default function UserLesson(props: IUserLessonProps): JSX.Element {
   return (
