+++ client/src/components/Dashboard/UserLessonList/UserLessonList.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import { IUserLesson, IUserLessonListProps } from '../../../interfaces';
 import { AppState } from '../../../store/configureStore';
-import UserLesson from '../UserLesson/UserLesson';
+import { UserLesson } from '../../../components';
 
 export default function UserLessonList(props: IUserLessonListProps): JSX.Element {
   const lesson = useSelector((state: AppState) => state.lesson.lesson);
