+++ client/src/components/Lesson/Terminal/Terminal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ITerminalResponse } from '../../../interfaces/lesson';
-import TerminalResponse from '../TerminalResponse/TerminalResponse';
+import { TerminalResponse } from '../../../components';
 
 export default function Terminal(props: { responses: ITerminalResponse[] }): JSX.Element {
   const responseList = props.responses.map((response) => <TerminalResponse response={response} />);
