+++ client/src/containers/Landing/Landing.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import pong from '../../animations/pong.json';
-import LottieAnimation from '../../components/LottieAnimation/LottieAnimation';
+import { LottieAnimation } from '../../components';
 
 export default function Landing(): JSX.Element {
   return (
