+++ client/src/containers/Lesson/Lesson.tsx
@@ -2,13 +2,10 @@ import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { AppState } from '../../store/configureStore';
 import { fetchLesson, fetchUserLessons } from '../../actions';
-import Instructions from '../../components/Lesson/Instructions/Instructions';
-import CodeEditor from '../../components/Lesson/CodeEditor/CodeEditor';
-import TaskList from '../../components/Lesson/TaskList/TaskList';
-import Terminal from '../../components/Lesson/Terminal/Terminal';
 import { validator } from '../../components/Lesson/Validation/validator';
 import { ITerminalResponse } from '../../interfaces/lesson';
 import { updateUserLessons } from '../../actions/userLessons';
+import { CodeEditor, Instructions, TaskList, Terminal } from '../../components';
 
 export default function Lesson(): JSX.Element {
   const dispatch = useDispatch();
